move table engine proxy to table_engine crate.
